+++ docs/source/guides/sampling/plot_sampler_5.py
@@ -8,7 +8,7 @@ from bokeh.layouts import layout
 from omegaconf import OmegaConf
 
 from torch_brain.data import Dataset
-from torch_brain.data.sampler import RandomFixedWindowSampler
+from torch_brain.sampler import RandomFixedWindowSampler
 
 from _utils import download_file_from_s3
