+++ tests/test_sampler.py
@@ -5,7 +5,7 @@ import torch
 
 from temporaldata import Interval
 
-from torch_brain.data.sampler import (
+from torch_brain.sampler import (
     SequentialFixedWindowSampler,
     RandomFixedWindowSampler,
     TrialSampler,
