+++ torch_brain/__init__.py
@@ -5,6 +5,7 @@ from . import models
 from . import utils
 from . import transforms
 from . import registry
+from . import sampler
 
 from .registry import register_modality, get_modality_by_id, MODALITY_REGISTRY
