+++ torch_brain/sampler/__init__.py
@@ -1,4 +1,10 @@
 from .random_fixed_window import RandomFixedWindowSampler
+from .sequential_fixed_window import SequentialFixedWindowSampler
+from .trial_sampler import TrialSampler
+from .distributed_sampler import (
+    DistributedEvaluationSamplerWrapper,
+    DistributedStitchingFixedWindowSampler,
+)
 
 __all__ = [
     "RandomFixedWindowSampler",
