+++ torch_brain/sampler/sequential_fixed_window.py
@@ -1,0 +1,97 @@
+import logging
+from typing import List, Dict, Tuple, Optional
+from functools import cached_property
+
+import torch
+
+
+from torch_brain.dataset import DatasetIndex
+
+
+class SequentialFixedWindowSampler(torch.utils.data.Sampler):
+    r"""Samples fixed-length windows sequentially, always in the same order. The
+    sampling intervals are defined in the :obj:`sampling_intervals` parameter.
+    :obj:`sampling_intervals` is a dictionary where the keys are the session ids and the
+    values are lists of tuples representing the start and end of the intervals
+    from which to sample.
+
+    If the length of a sequence is not evenly divisible by the step, the last
+    window will be added with an overlap with the previous window. This is to ensure
+    that the entire sequence is covered.
+
+    Args:
+        sampling_intervals (Dict[str, List[Tuple[float, float]]]): Sampling intervals for each
+            session in the dataset.
+        window_length (float): Length of the window to sample.
+        step (float, optional): Step size between windows. If None, it
+            defaults to ``window_length``.
+        drop_short (bool, optional): Whether to drop windows smaller than ``window_length``.
+            Defaults to False.
+    """
+
+    def __init__(
+        self,
+        *,
+        sampling_intervals: Dict[str, List[Tuple[float, float]]],
+        window_length: float,
+        step: Optional[float] = None,
+        drop_short=False,
+    ):
+        self.sampling_intervals = sampling_intervals
+        self.window_length = window_length
+        self.step = step or window_length
+        self.drop_short = drop_short
+
+        assert self.step > 0, "Step must be greater than 0."
+
+    # we cache the indices since they are deterministic
+    @cached_property
+    def _indices(self) -> List[DatasetIndex]:
+        indices = []
+        total_short_dropped = 0.0
+
+        for session_name, sampling_intervals in self.sampling_intervals.items():
+            for start, end in zip(sampling_intervals.start, sampling_intervals.end):
+                interval_length = end - start
+                if interval_length < self.window_length:
+                    if self.drop_short:
+                        total_short_dropped += interval_length
+                        continue
+                    else:
+                        raise ValueError(
+                            f"Interval {(start, end)} is too short to sample from. "
+                            f"Minimum length is {self.window_length}."
+                        )
+
+                indices_ = [
+                    DatasetIndex(
+                        session_name, t.item(), (t + self.window_length).item()
+                    )
+                    for t in torch.arange(start, end, self.step, dtype=torch.float64)
+                    if t + self.window_length <= end
+                ]
+
+                indices.extend(indices_)
+
+                # we need to make sure that the entire interval is covered
+                if indices_[-1].end < end:
+                    indices.append(
+                        DatasetIndex(session_name, end - self.window_length, end)
+                    )
+
+        if self.drop_short and total_short_dropped > 0:
+            num_samples = len(indices)
+            logging.warning(
+                f"Skipping {total_short_dropped} seconds of data due to short "
+                f"intervals. Remaining: {num_samples * self.window_length} seconds."
+            )
+            if num_samples == 0:
+                raise ValueError("All intervals are too short to sample from.")
+
+        return indices
+
+    def __len__(self):
+        return len(self._indices)
+
+    def __iter__(self):
+        yield from self._indices
