+++ torch_brain/sampler/trial_sampler.py
@@ -1,0 +1,50 @@
+from typing import List, Dict, Tuple, Optional
+
+import torch
+from torch_brain.dataset import DatasetIndex
+
+
+class TrialSampler(torch.utils.data.Sampler):
+    r"""Randomly samples a single trial interval from the given intervals.
+
+    Args:
+        sampling_intervals (Dict[str, List[Tuple[int, int]]]): Sampling intervals for each
+            session in the dataset.
+        generator (Optional[torch.Generator], optional): Generator for shuffling.
+            Defaults to None.
+        shuffle (bool, optional): Whether to shuffle the indices. Defaults to False.
+    """
+
+    def __init__(
+        self,
+        *,
+        sampling_intervals: Dict[str, List[Tuple[float, float]]],
+        generator: Optional[torch.Generator] = None,
+        shuffle: bool = False,
+    ):
+        self.sampling_intervals = sampling_intervals
+        self.generator = generator
+        self.shuffle = shuffle
+
+    def __len__(self):
+        return sum(len(intervals) for intervals in self.sampling_intervals.values())
+
+    def __iter__(self):
+        # Flatten the intervals from all sessions into a single list
+        all_intervals = [
+            (session_id, start, end)
+            for session_id, intervals in self.sampling_intervals.items()
+            for start, end in zip(intervals.start, intervals.end)
+        ]
+
+        indices = [
+            DatasetIndex(session_id, start, end)
+            for session_id, start, end in all_intervals
+        ]
+
+        if self.shuffle:
+            # Yield a single DatasetIndex representing the selected interval
+            for idx in torch.randperm(len(indices), generator=self.generator):
+                yield indices[idx]
+        else:
+            yield from indices
